Factor out common translations in Surveys frontend
