+++ client/app/bundles/course/surveys/translations.jsx
@@ -1,0 +1,26 @@
+import { defineMessages } from 'react-intl';
+
+const translations = defineMessages({
+  title: {
+    id: 'course.surveys.fields.title',
+    defaultMessage: 'Title',
+  },
+  description: {
+    id: 'course.surveys.fields.description',
+    defaultMessage: 'Description',
+  },
+  opensAt: {
+    id: 'course.surveys.fields.openAt',
+    defaultMessage: 'Opens At',
+  },
+  expiresAt: {
+    id: 'course.surveys.fields.expiresAt',
+    defaultMessage: 'Expires At',
+  },
+  points: {
+    id: 'course.surveys.fields.points',
+    defaultMessage: 'Points Awarded',
+  },
+});
+
+export default translations;
